revert to user context being set in index page to help with testing
